ardunio-backend/requirements.txt
Adafruit-Blinka==8.18.1
adafruit-circuitpython-busdevice==5.2.4
adafruit-circuitpython-dht==3.7.8
adafruit-circuitpython-requests==1.13.2
adafruit-circuitpython-typing==1.9.2
adafruit-io==2.7.0
Adafruit-PlatformDetect==3.45.1
Adafruit-PureIO==1.1.10
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.1
DateTime==5.1
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.92.0
filelock==3.12.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
paho-mqtt==1.6.1
pandas==1.5.2
passlib==1.7.4
platformdirs==3.5.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
pyftdi==0.54.0
PyJWT==2.6.0
pymongo==4.3.3
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
pytz==2022.7.1
pyusb==1.2.1
PyYAML==6.0
requests==2.29.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.25.0
typing_extensions==4.4.0
urllib3==1.26.15
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.23.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.3
zope.interface==6.0
